+++ cmd/server/database.go
@@ -72,6 +72,9 @@ func (a *Application) dbOpen() error {
 				zoom    real    not null,
 				location text   not null,
 				islocal integer(1) not null,
+				frames integer not null default 0,
+				speed integer not null default 0,
+				loops integer not null default 0,
 					primary key (name,zoom)
 		);`)
 
@@ -92,12 +95,26 @@ func (a *Application) dbClose() error {
 	return a.sqldb.Close()
 }
 
-func (a *Application) StoreImageData(imageName string, img mapper.ImageInstance) error {
-	result, err := a.sqldb.Exec(`REPLACE INTO images (name, zoom, location, islocal) VALUES (?, ?, ?, ?);`, imageName, img.Zoom, img.File, img.IsLocalFile)
-	if err != nil {
-		return err
+func (a *Application) StoreImageData(imageName string, img mapper.ImageInstance, anim *mapper.ImageAnimation) error {
+	if anim == nil {
+		result, err := a.sqldb.Exec(`REPLACE INTO images (name, zoom, location, islocal) VALUES (?, ?, ?, ?);`, imageName, img.Zoom, img.File, img.IsLocalFile)
+		if err != nil {
+			return err
+		}
+
+		a.debugDbAffected(result, fmt.Sprintf("stored image record \"%s\"@%v local=%v, ID=%v", imageName, img.Zoom, img.IsLocalFile, img.File))
+	} else {
+		result, err := a.sqldb.Exec(`REPLACE INTO images (name, zoom, location, islocal, frames, speed, loops) VALUES (?, ?, ?, ?, ?, ?, ?);`,
+			imageName, img.Zoom, img.File, img.IsLocalFile,
+			anim.Frames, anim.FrameSpeed, anim.Loops,
+		)
+		if err != nil {
+			return err
+		}
+
+		a.debugDbAffected(result, fmt.Sprintf("stored image record \"%s\"@%v local=%v, ID=%v, frames=%d, speed=%d mS/frame, loops=%d", imageName, img.Zoom, img.IsLocalFile, img.File, anim.Frames, anim.FrameSpeed, anim.Loops))
 	}
-	a.debugDbAffected(result, fmt.Sprintf("stored image record \"%s\"@%v local=%v, ID=%v", imageName, img.Zoom, img.IsLocalFile, img.File))
+
 	return nil
 }
 
@@ -135,24 +152,35 @@ func (a *Application) QueryImageData(img mapper.ImageDefinition) (mapper.ImageDe
 	var resultSet mapper.ImageDefinition
 
 	a.Debugf(DebugDB, "query of image \"%s\"", img.Name)
-	rows, err := a.sqldb.Query(`SELECT zoom, location, islocal FROM images WHERE name=?`, img.Name)
+	rows, err := a.sqldb.Query(`SELECT zoom, location, islocal, frames, speed, loops FROM images WHERE name=?`, img.Name)
 	if err != nil {
 		return resultSet, err
 	}
 	defer rows.Close()
 
 	resultSet.Name = img.Name
+
 	for rows.Next() {
 		var instance mapper.ImageInstance
 		var isLocal int
+		var aframes int
+		var aspeed int
+		var aloops int
 
-		if err := rows.Scan(&instance.Zoom, &instance.File, &isLocal); err != nil {
+		if err := rows.Scan(&instance.Zoom, &instance.File, &isLocal, &aframes, &aspeed, &aloops); err != nil {
 			return resultSet, err
 		}
 		if isLocal != 0 {
 			instance.IsLocalFile = true
 		}
 		resultSet.Sizes = append(resultSet.Sizes, instance)
+		if resultSet.Animation == nil && (aframes != 0 || aspeed != 0 || aloops != 0) {
+			resultSet.Animation = &mapper.ImageAnimation{
+				Frames:     aframes,
+				FrameSpeed: aspeed,
+				Loops:      aloops,
+			}
+		}
 		a.Debugf(DebugDB, "result: \"%s\"@%v from \"%s\" (local=%v)", img.Name, instance.Zoom, instance.File, instance.IsLocalFile)
 	}
 	return resultSet, rows.Err()